% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_drape_lines}
\alias{up_drape_lines}
\title{Drape Geometries Using Column Value}
\usage{
up_drape_lines(x, col)
}
\arguments{
\item{x}{An \code{sf} object.}

\item{col}{Name of column containing offset values.}
}
\value{
An \code{sf} object.
}
\description{
Drape Geometries Using Column Value
}
