% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_get_dims}
\alias{up_get_dims}
\title{Generate list of dimension-related quantities.}
\usage{
up_get_dims(df, n, type = "horizontal")
}
\arguments{
\item{df}{\code{sf} object.}

\item{n}{Number of regularly-spaced lines to generate.}

\item{type}{String indicating direction of lines---options are "horizontal" and "vertical".}
}
\value{
A list of interval-relevant values.
}
\description{
Generate list of dimension-related quantities.
}
