% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_interpolate_vector}
\alias{up_interpolate_vector}
\title{Interpolate polygons using inverse distance weighting}
\usage{
up_interpolate_vector(x, field, nmax = 15, maxdist = Inf, idp = 2)
}
\arguments{
\item{x}{An \code{sf} object containing POLYGONs.}

\item{field}{Character. Name of field whose values should be interpolated.}

\item{nmax}{Maximum number of adjacent observations.}

\item{maxdist}{Maximum distance to be counted.}

\item{idp}{Inverse distance power. Defaults to 2.}
}
\value{
A \code{SpatRaster}.
}
\description{
Interpolate polygons using inverse distance weighting
}
