% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_mask_lines}
\alias{up_mask_lines}
\title{Mask lines by shapes.}
\usage{
up_mask_lines(x, y)
}
\arguments{
\item{x}{An \code{sf} object to mask.}

\item{y}{An \code{sf} object to serve as mask/}
}
\value{
An \code{sf} object.
}
\description{
Mask lines by shapes.
}
