% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_offset_lines}
\alias{up_offset_lines}
\title{Offset Points based on a Column Value and an Angle}
\usage{
up_offset_lines(x, angle, col)
}
\arguments{
\item{x}{An \code{sf} object containing points.}

\item{angle}{Angle in degrees.}

\item{col}{Name of column containing offset values.}
}
\value{
\code{sf} object containing offset points.
}
\description{
Offset Points based on a Column Value and an Angle
}
