% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_regular_lines}
\alias{up_regular_lines}
\title{Generate \code{n} regularly-spaced lines over \code{df} extent.}
\usage{
up_regular_lines(df, dims, mask = TRUE)
}
\arguments{
\item{df}{An \code{sf} dataframe.}

\item{dims}{Object returned by \code{up_get_dims()}.}

\item{mask}{Whether resulting lines should be clipped to df extent.}
}
\value{
A sf dataframe of regularly spaced POLYLINEs.
}
\description{
Generate \code{n} regularly-spaced lines over \code{df} extent.
}
