% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_rotate}
\alias{up_rotate}
\title{Construct a Rotation Matrix.}
\usage{
up_rotate(a)
}
\arguments{
\item{a}{Angle in radians.}
}
\value{
A rotation matrix.
}
\description{
Construct a Rotation Matrix.
}
