% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_rotate_extent}
\alias{up_rotate_extent}
\title{Rotate a Given \code{sf} Object by an Arbitrary Angle.}
\usage{
up_rotate_extent(x, angle)
}
\arguments{
\item{x}{An \code{sf} object.}

\item{angle}{Angle in degrees.}
}
\value{
A rotated \code{sf} object.
}
\description{
Rotate a Given \code{sf} Object by an Arbitrary Angle.
}
