% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_sample_lines}
\alias{up_sample_lines}
\title{Sample Along Lines}
\usage{
up_sample_lines(x, interval)
}
\arguments{
\item{x}{An \code{sf} object containing lines.}

\item{interval}{Linear units. Interval at which to sample lines.}
}
\value{
An \code{sf} object containing sampled points.
}
\description{
Sample Along Lines
}
