% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_unknown_pleasures}
\alias{up_unknown_pleasures}
\title{Transforms regularly-spaced lines into an "Unknown Pleasures"-esque set of regular section cuts based on raster value.}
\usage{
up_unknown_pleasures(
  lines,
  raster,
  dims,
  interval,
  max = NULL,
  bleed_factor = 1.5,
  mode = "planar",
  polygon = TRUE
)
}
\arguments{
\item{lines}{\code{sf} object containing regularly spaced lines.}

\item{raster}{\code{raster} object.}

\item{dims}{An object returned by the \code{up_get_dims()} function.}

\item{interval}{Interval. Can be \code{units} object. Otherwise will be in unit of CRS.}

\item{max}{Numeric.}

\item{bleed_factor}{How much should maxima bleed into the area of the line above or below.}

\item{mode}{If \code{planar}, results will be planar offset lines. If \code{xyz}, lines will be offset on \code{LINESTRING} z axis.}

\item{polygon}{If \code{TRUE}, outputs polygons (or closed linestrings if paired with \code{mode = "xyz"}. If \code{FALSE}, outputs lines with no baseline.}
}
\value{
\code{sf} object  containing "Unknown Pleasures" features.
}
\description{
Transforms regularly-spaced lines into an "Unknown Pleasures"-esque set of regular section cuts based on raster value.
}
