% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownpleasur.R
\name{up_regular_lines}
\alias{up_regular_lines}
\title{Generate regularly-spaced lines an an angle over an extent.}
\usage{
up_regular_lines(x, cellsize, angle = 0)
}
\arguments{
\item{x}{An \code{sf} dataframe.}

\item{cellsize}{A spacing distance in linear units.}

\item{angle}{Angle in degrees. 0 is horizontal.}
}
\value{
\code{sf} object of regularly spaced \code{LINESTRING}s.
}
\description{
Generate regularly-spaced lines an an angle over an extent.
}
